Move housing packets into Network namespace
